WIP
Changes:
- Modificada a forma de criação do grafo estava a matriz nao estava a ser iniciada corretamente.
- Adicionado player à classe divisao
- Criada ADT para Jogo e Inicio do Controlador de Jogo
- Movimentação de Player e Inimigo funcional
- Inimigo atualiza o peso das arestas adjacentes ao grafo que ele se move

+++ src/main/java/lei/estg/models/Interfaces/EdificioADT.java
@@ -10,5 +10,5 @@ public interface EdificioADT<T> extends NetworkADT<T> {
     public Iterator<T> getVertex();
     public Iterator<T> getAdjacentes(T divisao);
     public Iterator<T> encontrarCaminho(T divisao1, T divisao2) throws EmptyStackException;
-    public void updateEdge(int index1, int index2, int weight);
+    public void updateEdge(int index1, int index2, double weight);
 }
